requirements.txt
Flask
Werkzeug==2.0.3
Flask-Bcrypt==0.7.1
python-dotenv==0.19.0
psycopg2-binary